fix: card storybook
