+++ src/components/Card/Card.module.scss
@@ -1,8 +1,6 @@
 @import '@/styles/variables.scss';
 
 .card {
-  display: flex;
-  flex-direction: column;
   cursor: pointer;
   padding: 20px 10px 0 10px;
   font-size: 1em;
